+++ BeatSpy/Commands/CommandBase.cs
@@ -1,16 +1,13 @@
 ﻿using System;
 using System.Windows.Input;
 
-namespace BeatSpy.Commands.Base;
+namespace BeatSpy.Commands;
 
 internal abstract class CommandBase : ICommand
 {
     public event EventHandler? CanExecuteChanged;
 
-    public virtual bool CanExecute(object? parameter)
-    {
-        return true;
-    }
+    public virtual bool CanExecute(object? parameter) { return true; }
 
     public abstract void Execute(object? parameter);
